+++ my-app/src/app/components/LoginForm.tsx
@@ -27,7 +27,7 @@ export default function LoginForm() {
   }
   return (
     <div className="flex flex-col">
-      <p className="mb-8 subtitle-1 text-color-neutral-110">
+      <p className="my-8 subtitle-1 text-color-neutral-110">
         Faça login com email
       </p>
       <ThemeProvider theme={TextFieldTheme}>
@@ -46,7 +46,7 @@ export default function LoginForm() {
             id="outlined-adornment-password"
             type={showPassword ? 'text' : 'password'}
             endAdornment={
-              <InputAdornment position="end">
+              <InputAdornment position="start">
                 <IconButton
                   aria-label="toggle password visibility"
                   onClick={handleClickShowPassword}
@@ -68,11 +68,10 @@ export default function LoginForm() {
         size="large"
         disabled={false}
         name="ENTRAR"
-        className="mb-[1.13rem]"
       />
       <a
         href="../register"
-        className="subtitle-1 text-color-neutral-100 !no-underline"
+        className="subtitle-1 mt-[1.13rem] text-color-neutral-100 !no-underline"
       >
         Cadastre-se
       </a>
