+++ my-app/src/app/layout.tsx
@@ -1,4 +1,5 @@
 import type { Metadata } from 'next'
+import { GoogleOAuthProvider } from '@react-oauth/google'
 import '@fontsource/roboto/300.css'
 import '@fontsource/roboto/400.css'
 import '@fontsource/roboto/500.css'
@@ -11,14 +12,20 @@ export const metadata: Metadata = {
   description: 'Generated by create next app'
 }
 
+// Google Client Id
+const ClientId =
+  '246885474721-45d51ac88psutifbi0mnf7ssqv7j3tpb.apps.googleusercontent.com'
+
 export default function RootLayout({
   children
 }: Readonly<{
   children: React.ReactNode
 }>) {
   return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
+    <GoogleOAuthProvider clientId={ClientId}>
+      <html lang="en">
+        <body>{children}</body>
+      </html>
+    </GoogleOAuthProvider>
   )
 }
