+++ my-app/src/app/lib/definiton.ts
@@ -4,10 +4,5 @@ export type TypeButton = {
   color: string
   size: string
   disabled: boolean
-  className: string
   name: string
-}
-
-export type GoogleLogin = {
-  className: string
 }
