+++ my-app/src/app/themes/TextField.ts
@@ -4,8 +4,7 @@ import { createTheme } from '@mui/material/styles'
 export const TextFieldTheme = createTheme({
   palette: {
     primary: {
-      main: '#2348B1',
-      contrastText: '#000000'
+      main: '#2348B1'
     }
   }
 })
